% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{accuracy}
\usage{
accuracy(signals)
}
\arguments{
\item{signals}{output of the function signalCalculation. A dataFrame including the original point data and the signals sent by the reference regions.}
}
\value{
The dataFrame with the species occurrence information and an extra column indicating the estimated accuracy of each point.
}
\description{
Estimates the accuracy of each point record.
}
