% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableCountries.R
\name{availableCountries}
\alias{availableCountries}
\title{availableCountries}
\usage{
availableCountries()
}
\value{
This function provides a list of countries and entities names 
available with rworldmaps for checklists
}
\description{
List of countries and entities names for checklists
}
\examples{
country_list <- availableCountries()
}
