% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogeoStatus.R
\name{biogeoStatus}
\alias{biogeoStatus}
\title{biogeoStatus}
\usage{
biogeoStatus(signals)
}
\arguments{
\item{signals}{output of the function signalCalculation. A dataFrame 
including the original point data and the signals sent by the reference 
regions.}
}
\value{
The dataFrame with the species occurrence information and an extra 
column indicating the estimated biogeographic status of each point.
}
\description{
Estimates the biogeographic status of each point record.
}
