% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryChecklists.R
\name{countryChecklist}
\alias{countryChecklist}
\title{countryChecklists}
\usage{
countryChecklist(countries, biogeo_status)
}
\arguments{
\item{biogeo_status}{vector informing the status of each country: alien, native or unknown}

\item{country}{vector with one or more country names}
}
\value{
This function provides shapefiles of countries with the correspondent biogeographic status of the species.
}
\description{
Prepares user provided reference regions on a country level
}
\examples{
country_checklist <- countryChecklist(c("Brazil","Argentina","Uruguay","Paraguay"),
                                      c("native","alien","unknown","native"))
}
