% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGbifDecade.R
\name{getGbifDecade}
\alias{getGbifDecade}
\title{Download gbif data per decade}
\usage{
getGbifDecade(species)
}
\arguments{
\item{species}{character, species binomial name}
}
\value{
The function downloads gbif records decade by decade for a given species in a given period of time. It is a secondary function to cope with the limitation of 200,000 records that can be downloaded from a single request through rgbif package. This function is composed of two parts: 1 downloads all occurrences for a species from 1001 to 1900, 2 loops thought decades from 1901 to present day
}
\description{
Downloads gbif records decade by decade
}
