% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGbifMonth.R
\name{getGbifMonth}
\alias{getGbifMonth}
\title{Download gbif data per month}
\usage{
getGbifMonth(species, year)
}
\arguments{
\item{species}{character, species binomial name}

\item{year}{numeric, the year}
}
\value{
The function downloads gbif records month by month for a given species in a given year. It is a secondary function to cope with the limitation of 200,000 records that can be downloaded from a single request through rgbif package.
}
\description{
Downloads gbif records month by month
}
