% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGbifYear.R
\name{getGbifYear}
\alias{getGbifYear}
\title{Download gbif data per year}
\usage{
getGbifYear(species, year_0, year_1)
}
\arguments{
\item{species}{character, species binomial name}

\item{year_0}{numeric, the first year of the period}

\item{year_1}{numeric, the last year of the period}
}
\value{
The function downloads gbif records year by year for a given species in a given period of time. It is a secondary function to cope with the limitation of 200,000 records that can be downloaded from a single request through rgbif package.
}
\description{
Downloads gbif records year by year
}
