% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftRegions.R
\name{giftRegions}
\alias{giftRegions}
\title{giftRegions}
\usage{
giftRegions(species)
}
\arguments{
\item{species}{character, species binomial name}
}
\value{
This function returns a list containing three shapefiles derived by information supplied by GIFT. "regs" includes all the features corresponding to regions where the species has been listed as present. "regs_native" includes all the features corresponding to regions where the species has been listed as native. And "regs_alien" includes all the features corresponding to regions where the species has been listed as alien.
}
\description{
Gets regions listed by GIFT for plant species
}
\examples{
gift_reference_regions <- giftRegions("Boreava aptera")
}
