% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRacatus-package.R
\docType{data}
\name{H_mirandae_sp}
\alias{H_mirandae_sp}
\title{Hemitriccus mirandae spatialPoints exemplary file}
\description{
A \code{SpatialPointsDataFrame} containing the occurrences of 
 Hemitriccus mirandae downloaded from GBIF
}
\examples{

data("H_mirandae_sp")

}
\keyword{examplary}
\keyword{object}
