% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRacatus-package.R
\docType{data}
\name{Model_biogeo}
\alias{Model_biogeo}
\title{Biogeographical model}
\description{
A \code{glm} biogeographical model
}
\examples{

data("Model_biogeo")

}
\keyword{model}
\keyword{object}
