% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occID.R
\name{occID}
\alias{occID}
\title{occID}
\usage{
occID(occ_sp)
}
\arguments{
\item{occ_sp}{spatialPointsDataTable of the species occurrence.}
}
\value{
A dataFrame including the original data in the input and the cellID of each point record
}
\description{
Extracts cellIDs of presence locations
}
\examples{
occ_sp <- bRacatus::H_mirandae_sp
occ_ID <- occID(occ_sp)

}
