% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occSpatialPoints.R
\name{occSpatialPoints}
\alias{occSpatialPoints}
\title{occSpatialPoints}
\usage{
occSpatialPoints(occ)
}
\arguments{
\item{occ}{table}
}
\value{
This function used the output of getOcc or giveOcc to create spatialPoints from tables.
}
\description{
Creates spatialpoints from tables containing species names and coordinates.
}
