% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAccuracy.R
\name{plotAccuracy}
\alias{plotAccuracy}
\title{plotAccuracy}
\usage{
plotAccuracy(acc, regional = TRUE)
}
\arguments{
\item{acc}{dataTable of the species occurrence including a column with the estimated accuracy of points.}

\item{regional}{logical, whether the whole world should be plotted as the background or only the region adjacent to the species countries of occurrence.}
}
\value{
This function plots the species occurrence with estimated accuracy of all points.
}
\description{
Plot the species occurrences showing the estimated accuracy of points.
}
