% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiogeoStatus.R
\name{plotBiogeoStatus}
\alias{plotBiogeoStatus}
\title{plotAccuracy}
\usage{
plotBiogeoStatus(biogeo, regional = TRUE)
}
\arguments{
\item{regional}{logical, whether the whole world should be plotted as the background or only the region adjacent to the species countries of occurrence.}

\item{acc}{dataTable of the species occurrence including a column with the estimated biogeographical status of points.}
}
\value{
This function plots the species occurrence with estimated biogeographical status of all points.
}
\description{
Plot the species occurrences showing the estimated biogeographical status of points.
}
