% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRacatus-package.R
\docType{data}
\name{Range_Phalanger_orientalis}
\alias{Range_Phalanger_orientalis}
\title{Range Phalanger orientalis}
\description{
A \code{SpatialPolygonsDataFrame} Range Phalanger orientalis
}
\examples{

data("Range_Phalanger_orientalis")

}
\keyword{examplary}
\keyword{object}
