% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRacatus-package.R
\docType{data}
\name{signals}
\alias{signals}
\title{signals example 1}
\description{
A \code{SpatialPolygonsDataFrame} signals example 1
}
\examples{

data("signals")

}
\keyword{examplary}
\keyword{object}
