% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRacatus-package.R
\docType{data}
\name{signals_2}
\alias{signals_2}
\title{signals example 2}
\description{
A \code{SpatialPolygonsDataFrame} signals example 2
}
\examples{

data("signals_2")

}
\keyword{examplary}
\keyword{object}
