% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRacatus-package.R
\docType{data}
\name{signals_3}
\alias{signals_3}
\title{signals example 3}
\description{
A \code{SpatialPolygonsDataFrame} signals example 3
}
\examples{

data("signals_3")

}
\keyword{examplary}
\keyword{object}
