% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bRacatus-package.R
\docType{data}
\name{ID_raster}
\alias{ID_raster}
\title{Cell ID raster}
\description{
A \code{Raster} half-degree raster of the world with unique IDs per cell
}
\examples{

data("ID_raster")

}
\keyword{reference}
