% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occID.R
\name{occID}
\alias{occID}
\title{occID}
\usage{
occID(occ)
}
\arguments{
\item{occ}{dataTable of the species occurrence.}
}
\value{
A dataFrame including the original data in the input and the 
cellID of each point record
}
\description{
Extracts cellIDs of presence locations
}
\examples{
occ <- getOcc("Hemitriccus mirandae")
occ_ID <- occID(occ)

}
