% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occID.R
\name{occID}
\alias{occID}
\title{occID}
\usage{
occID(occ_sp)
}
\arguments{
\item{occ_sp}{spatialPointsDataTable of the species occurrence.}
}
\value{
A spatialPointsDataTable including the cellID of each point record
}
\description{
Extracts cellIDs of presence locations
}
\examples{
country_checklist <- countryChecklist(c("Brazil","Argentina","Uruguay","Paraguay"),
c("native","alien","unknown","native"))

rasterised_checklist <- rasteriseChecklists(country_checklist)

value_IDs <- valueID(rasterised_checklist)
}
