% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occSpatialPoints.R
\name{occSpatialPoints}
\alias{occSpatialPoints}
\title{occSpatialPoints}
\usage{
occSpatialPoints(occ)
}
\arguments{
\item{occ}{table}
}
\value{
This function creates spatialpoints from tables containing coordinates.
}
\description{
Downloads gbif records iterating when necessary to overcome the limitation of 200,000 records
}
\examples{
# Create a data.frame containing species names and coordinates

sps_occurrence <- getOcc("Hemitriccus mirandae")
sps_sp <- occSpatialPoints(sps_occurrence)
}
