% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalCalculation.R
\name{signalCalculation}
\alias{signalCalculation}
\title{signalCalculation}
\usage{
signalCalculation(ref_reg, pts, biogeo = TRUE)
}
\arguments{
\item{ref_reg}{a list of shapefiles containing checklist regions as 
"presence", "native", and "alien" categories. These can be original 
checklists, or checklists derived from range maps.}

\item{pts}{data.frame containing the point records and their coordinates.}

\item{biogeo}{logical, whether the biogeographical status indices should 
be calculated or not. Default is true, however at least the native 
reference regions must be included in the data.}
}
\value{
The data.frame of species occurrences with extra columns containing
the location ID and presence signals for each point. If biogeo=TRUE, the 
data.frame also includes the nativeness and alienness indices.
}
\description{
Calculates signals sent from reference regions to point records.
}
