% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueID.R
\name{valueID}
\alias{valueID}
\title{valueID}
\usage{
valueID(checklists_raster)
}
\arguments{
\item{checklists_raster}{List containing rasterised checklists for presence, native and alien reference regions}
}
\value{
A list with cell IDs and signal values for all cells where the species is present, native and alien.
}
\description{
Extracts signal values and ID from each cell
}
\examples{
country_checklist <- countryChecklist(c("Brazil","Argentina","Uruguay","Paraguay"),
c("native","alien","unknown","native"))

rasterised_checklist <- rasteriseChecklists(country_checklist)

value_IDs <- valueID(rasterised_checklist)
}
